0.2.2025.1.12.1
Fixed bugs regarding Newtonsoft.Json by downgrading it to 12.x as the new versions were giving problems.
Added new functionalities to the system, now it's possible to save exceptions on the database so that it might be easier to check later.
Changed the way that the steps are saved, now the steps have their own object instead of being saved in a string as a json object or as pure string